feat(benefits-deductions): add deduction service to manage request for deductions calculations
