+++ backend/Application/DTOs/CalculateDeductionDto.cs
@@ -3,7 +3,7 @@ using backend.Domain;
 
 public class CalculateDeductionDto
 {
+    public Guid EmployeeId { get; set; }
     public CalculateGrossPaymentDto GrossPayment { get; set; }
-    public List<Benefit> Benefits { get; set; }
     public Guid PaymentDetailsId { get; set; }
 }
